removed unused files

+++ workflow/rules/exp_type_4.smk
@@ -46,16 +46,17 @@ if exp_type == 4:
             list_of_files = [dir_prefix + x for x in os.listdir(dir_prefix)]
 
             pivot = random.choice(list_of_files)
-            if not os.path.isdir(f"input_type4/pivot/dataset_{i}"):
-                os.makedirs(f"input_type4/pivot/dataset_{i}")
+            if not os.path.isdir(f"input_type4/pivot/"):
+                os.makedirs(f"input_type4/pivot/")
             
             # Remove pivot from other set of genomes
-            shutil.copy(pivot, f"input_type4/pivot/dataset_{i}/pivot_{i}.fna.gz")
-            os.remove(pivot)
+            shutil.copy(pivot, f"input_type4/pivot/pivot_{i}.fna.gz")
+            if(out_pivot):
+                os.remove(pivot)
     
     # Initialize complex ops directory to start writing files
-    if not os.path.isdir("complex_ops"):
-        os.makedirs("complex_ops")
+    if not os.path.isdir("complex_ops_type_4"):
+        os.makedirs("complex_ops_type_4")
 
     # Build the complex ops files
     for k in k_values:
@@ -100,11 +101,6 @@ if exp_type == 4:
         with open(f"filelists_type_4/k_{k}/intersections_filelist.txt", "w") as fd:
             for f in intersect_files:
                 fd.write(base_dir+f+"\n")
-
-    # Declare confusion matrix and accuracy files
-    if not os.path.isdir("accuracies"):
-        os.makedirs("accuracies_type_4/confusion_matrix")
-        os.makedirs("accuracies_type_4/accuracy")
     
 
 ####################################################
@@ -144,34 +140,37 @@ rule build_kmc_database_on_genome_exp_type_4:
     input:
         "input_type4/rest_of_set/dataset_{num}/{genome}.fna.gz"
     output:
-        "step_1_type4/rest_of_set/k_{k}/dataset_{num}/{genome}.kmc_pre",
-        "step_1_type4/rest_of_set/k_{k}/dataset_{num}/{genome}.kmc_suf"
+        "step_1_type_4/rest_of_set/k_{k}/dataset_{num}/{genome}.kmc_pre",
+        "step_1_type_4/rest_of_set/k_{k}/dataset_{num}/{genome}.kmc_suf"
     shell:
-        "kmc -fm -m64 -k{wildcards.k} -ci1 {input} step_1_type4/rest_of_set/k_{wildcards.k}/dataset_{wildcards.num}/{wildcards.genome} tmp/"
+        "kmc -fm -m64 -k{wildcards.k} -ci1 {input} step_1_type_4/rest_of_set/k_{wildcards.k}/dataset_{wildcards.num}/{wildcards.genome} tmp/"
 
 rule build_kmc_database_on_pivot_exp_type_4:
     input:
-        "input_type4/pivot/dataset_{num}/pivot_{num}.fna.gz"
+        "input_type4/pivot/pivot_{num}.fna.gz"
     output:
-        "step_1_type4/pivot/k_{k}/dataset_{num}/pivot_{num}.kmc_pre",
-        "step_1_type4/pivot/k_{k}/dataset_{num}/pivot_{num}.kmc_suf"
+        "step_1_type_4/pivot/k_{k}/dataset_{num}/pivot_{num}.kmc_pre",
+        "step_1_type_4/pivot/k_{k}/dataset_{num}/pivot_{num}.kmc_suf"
     shell:
-        "kmc -fm -m64 -k{wildcards.k} -ci1 {input} step_1_type4/pivot/k_{wildcards.k}/dataset_{wildcards.num}/pivot_{wildcards.num} tmp/"
+        "kmc -fm -m64 -k{wildcards.k} -ci1 {input} step_1_type_4/pivot/k_{wildcards.k}/dataset_{wildcards.num}/pivot_{wildcards.num} tmp/"
 
 # Section 3.2: Converts each kmer database into 
-# a set (multiplicity=1) for all databases.
+# a set (multiplicity=1) for all databases._
 
 rule transform_genome_to_set_exp_type_4:
     input:
-        "step_1_type4/rest_of_set/k_{k}/dataset_{num}/{genome}.kmc_pre",
-        "step_1_type4/rest_of_set/k_{k}/dataset_{num}/{genome}.kmc_suf"
+        "step_1_type_4/rest_of_set/k_{k}/dataset_{num}/{genome}.kmc_pre",
+        "step_1_type_4/rest_of_set/k_{k}/dataset_{num}/{genome}.kmc_suf"
     output:
         "genome_sets_type_4/rest_of_set/k_{k}/dataset_{num}/{genome}.transformed.kmc_pre",
         "genome_sets_type_4/rest_of_set/k_{k}/dataset_{num}/{genome}.transformed.kmc_suf"
     shell:
         """
-        kmc_tools transform step_1_type4/rest_of_set/k_{wildcards.k}/dataset_{wildcards.num}/{wildcards.genome} \
+        kmc_tools transform step_1_type_4/rest_of_set/k_{wildcards.k}/dataset_{wildcards.num}/{wildcards.genome} \
         set_counts 1 genome_sets_type_4/rest_of_set/k_{wildcards.k}/dataset_{wildcards.num}/{wildcards.genome}.transformed
+
+        # Remove step_1 rest of set
+        rm {input[0]} {input[1]}
         """
 
 # Section 3.3: Takes all genomes within a dataset
@@ -210,6 +209,9 @@ rule transform_union_to_set_exp_type_4:
      """
         kmc_tools transform unions_type_4/rest_of_set/k_{wildcards.k}/dataset_{wildcards.num}/dataset_{wildcards.num}.transformed.combined \
         set_counts 1 genome_sets_type_4/unions_type_4/k_{wildcards.k}/dataset_{wildcards.num}/dataset_{wildcards.num}.transformed.combined.transformed
+
+        # Remove unions rest of set
+        rm {input[0]} {input[1]}
      """
 
 # Section 3.4: Performs intersection
@@ -217,7 +219,7 @@ rule transform_union_to_set_exp_type_4:
 
 rule pivot_intersect_exp_type_4:
     input:
-        "step_1_type4/pivot/k_{k}/dataset_{pivot_num}/pivot_{pivot_num}.kmc_pre",
+        "step_1_type_4/pivot/k_{k}/dataset_{pivot_num}/pivot_{pivot_num}.kmc_pre",
         "genome_sets_type_4/unions_type_4/k_{k}/dataset_{num}/dataset_{num}.transformed.combined.transformed.kmc_pre"
     output:
         "intersection_results_type_4/k_{k}/pivot_{pivot_num}/pivot_{pivot_num}_intersect_dataset_{num}.kmc_pre",
@@ -225,7 +227,7 @@ rule pivot_intersect_exp_type_4:
     shell:
         """
         kmc_tools simple genome_sets_type_4/unions_type_4/k_{wildcards.k}/dataset_{wildcards.num}/dataset_{wildcards.num}.transformed.combined.transformed \
-        step_1_type4/pivot/k_{wildcards.k}/dataset_{wildcards.pivot_num}/pivot_{wildcards.pivot_num} intersect \
+        step_1_type_4/pivot/k_{wildcards.k}/dataset_{wildcards.pivot_num}/pivot_{wildcards.pivot_num} intersect \
         intersection_results_type_4/k_{wildcards.k}/pivot_{wildcards.pivot_num}/pivot_{wildcards.pivot_num}_intersect_dataset_{wildcards.num} -ocsum
         """
 
@@ -247,14 +249,14 @@ rule intersection_histogram_exp_type_4:
 
 rule pivot_text_dump_exp_type_4:
     input:
-        "step_1_type4/pivot/k_{k}/dataset_{num}/pivot_{num}.kmc_pre",
-        "step_1_type4/pivot/k_{k}/dataset_{num}/pivot_{num}.kmc_suf"
+        "step_1_type_4/pivot/k_{k}/dataset_{num}/pivot_{num}.kmc_pre",
+        "step_1_type_4/pivot/k_{k}/dataset_{num}/pivot_{num}.kmc_suf"
     output:
         "text_dump_type_4/k_{k}/pivot/pivot_{num}.txt"
     shell: 
         """
         kmc_tools transform \
-        step_1_type4/pivot/k_{wildcards.k}/dataset_{wildcards.num}/pivot_{wildcards.num} \
+        step_1_type_4/pivot/k_{wildcards.k}/dataset_{wildcards.num}/pivot_{wildcards.num} \
         dump -s text_dump_type_4/k_{wildcards.k}/pivot/pivot_{wildcards.num}.txt \
         """
 
@@ -278,8 +280,8 @@ rule run_merge_list_exp_type_4:
     input:
         get_filelists_and_text_dumps_exp_4
     output:
-        "accuracies_type_4/accuracy/k_{k}_accuracy.csv",
-        "accuracies_type_4/confusion_matrix/k_{k}_confusion_matrix.csv"
+        "accuracies_type_4/values/k_{k}_accuracy_values.csv",
+        "accuracies_type_4/confusion_matrix/k_{k}_confusion_matrix.txt"
     shell:
         """
         python3 {repo_dir}/src/merge_lists.py \
@@ -288,6 +290,10 @@ rule run_merge_list_exp_type_4:
         -o {base_dir}/accuracies_type_4/ \
         -n {num_datasets} \
         -k {wildcards.k} \
+
+        # Remove text dumps for this k
+        rm text_dump_type_4/k_{wildcards.k}/intersection/pivot_*/pivot_*_intersect_dataset_*.txt
+        rm text_dump_type_4/k_{wildcards.k}/pivot/pivot_*.txt
         """
 
 # Section 3.8 Concatonates accuracy score tables
@@ -295,8 +301,8 @@ rule run_merge_list_exp_type_4:
 
 rule concatonate_accuracies_exp_type4:
     input:
-        expand("accuracies_type_4/accuracy/k_{k}_accuracy.csv", k = k_values)
+        expand("accuracies_type_4/values/k_{k}_accuracy_values.csv", k = k_values)
     output:
-        "accuracies_type_4/accuracy_scores.csv"
+        "accuracies_type_4/accuracy_values.csv"
     shell:
-        "cat accuracies_type_4/accuracy/k_*_accuracy.csv > accuracies_type_4/accuracy_scores.csv"
+        "cat accuracies_type_4/values/*.csv > accuracies_type_4/accuracy_values.csv"
